Remove EasyDeploy component and related code
