+++ extensions/vscode/webviews/homeView/src/App.vue
@@ -3,7 +3,6 @@
 <template>
   <main>
     <EvenEasierDeploy class="easy-deploy-container" />
-    <EasyDeploy class="easy-deploy-container" />
     <template v-if="home.selectedConfiguration">
       <ProjectFiles v-model:expanded="projectFilesExpanded" />
       <PythonPackages />
@@ -15,7 +14,6 @@
 import { ref } from "vue";
 
 import EvenEasierDeploy from "src/components/EvenEasierDeploy.vue";
-import EasyDeploy from "src/components/EasyDeploy.vue";
 import ProjectFiles from "src/components/views/ProjectFiles.vue";
 import PythonPackages from "src/components/views/PythonPackages.vue";
