+++ extensions/vscode/webviews/homeView/src/components/EvenEasierDeploy.vue
@@ -35,14 +35,6 @@
       />
     </div>
 
-    <!-- Temporary Section -->
-    <p>Deployment: {{ home.selectedDeployment?.saveName || "undefined" }}</p>
-    <p>
-      Config: {{ home.selectedConfiguration?.configurationName || "undefined" }}
-    </p>
-    <p>Cred: {{ home.serverCredential?.name || " undefined" }}</p>
-    <!-- Remove this later -->
-
     <p v-if="home.selectedDeployment && !home.selectedConfiguration">
       The last Configuration used for this Destination was not found. Choose a
       new Configuration.
@@ -52,18 +44,75 @@
       A Credential for the Destination's server URL was not found. Create a new
       Credential.
     </p>
+
+    <DeployButton class="w-full" />
+
+    <template
+      v-if="home.selectedDeployment && home.selectedDeployment.serverType"
+    >
+      <vscode-divider class="home-view-divider" />
+
+      <div v-if="home.publishInProgress" class="progress-container">
+        <vscode-progress-ring class="progress-ring" />
+        Deployment in Progress...
+      </div>
+      <div v-else>
+        <h4 class="deployment-summary">
+          {{ lastStatusDescription }}
+        </h4>
+        <div
+          v-if="!isPreDeployment(home.selectedDeployment)"
+          class="last-deployment-time"
+        >
+          {{ formatDateString(home.selectedDeployment.deployedAt) }}
+        </div>
+        <div
+          v-if="home.selectedDeployment.deploymentError"
+          class="last-deployment-details last-deployment-error"
+        >
+          <span class="codicon codicon-error error-icon"></span>
+          <span class="error-message">
+            Error: {{ home.selectedDeployment.deploymentError.msg }}
+          </span>
+        </div>
+        <div class="last-deployment-details">
+          Targeting Posit Connect server at
+          <a
+            href=""
+            @click="navigateToUrl(home.selectedDeployment.serverUrl)"
+            >{{ home.selectedDeployment.serverUrl }}</a
+          >
+        </div>
+
+        <div
+          v-if="!isPreDeployment(home.selectedDeployment)"
+          class="last-deployment-details"
+        >
+          <vscode-button
+            appearance="secondary"
+            @click="navigateToUrl(home.selectedDeployment.dashboardUrl)"
+            class="w-full"
+          >
+            View Content
+          </vscode-button>
+        </div>
+      </div>
+    </template>
   </div>
 </template>
 
 <script setup lang="ts">
 import { computed } from "vue";
 
+import { isPreDeployment } from "../../../../src/api";
 import { WebviewToHostMessageType } from "../../../../src/types/messages/webviewToHostMessages";
 
 import { useHostConduitService } from "src/HostConduitService";
 import { useHomeStore } from "src/stores/home";
 import QuickPickItem from "src/components/QuickPickItem.vue";
 import ActionToolbar from "src/components/ActionToolbar.vue";
+import DeployButton from "src/components/DeployButton.vue";
+import { formatDateString } from "src/utils/date";
 
 const home = useHomeStore();
 const hostConduit = useHostConduitService();
@@ -95,6 +144,28 @@ const toolbarActions = computed(() => {
 const onSelectDestination = () => {
   hostConduit.sendMsg({
     kind: WebviewToHostMessageType.SELECT_DESTINATION,
+  });
+};
+
+const lastStatusDescription = computed(() => {
+  if (!home.selectedDeployment) {
+    return undefined;
+  }
+  if (home.selectedDeployment.deploymentError) {
+    return "Last Deployment Failed";
+  }
+  if (isPreDeployment(home.selectedDeployment)) {
+    return "Not Yet Deployed";
+  }
+  return "Last Deployment Successful";
+});
+
+const navigateToUrl = (url: string) => {
+  hostConduit.sendMsg({
+    kind: WebviewToHostMessageType.NAVIGATE,
+    content: {
+      uriPath: url,
+    },
   });
 };
 </script>
@@ -140,5 +211,48 @@ const onSelectDestination = () => {
 
 :deep(.action-item) {
   margin-right: 4px;
+}
+
+.home-view-divider {
+  margin-top: 1.33em;
+}
+
+.deployment-summary {
+  margin-bottom: 5px;
+}
+
+.last-deployment-time {
+  margin-bottom: 20px;
+}
+
+.last-deployment-details {
+  margin-top: 10px;
+}
+
+.last-deployment-error {
+  border: solid 2px;
+  border-color: gray;
+  padding: 5px;
+  display: flex;
+  align-items: center;
+}
+
+.error-icon {
+  flex: 0;
+}
+
+.error-message {
+  margin-left: 5px;
+}
+
+.progress-container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-top: 10px;
+}
+
+.progress-ring {
+  margin-right: 10px;
 }
 </style>
